requirements.txt
asgiref==3.7.2
Django==4.2.5 # Bumped due to vulnerability issues
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
python-decouple==3.8
psycopg2-binary==2.9.9
sqlparse==0.4.4 # Bumped due to vulnerability issues
